# .agents/resources/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "GraphState": {
        "description": "Complete graph state with nodes and edges.",
        "properties": {
          "edges": {
            "items": {
              "discriminator": {
                "mapping": {
                  "column_ref": "#/components/schemas/ColumnRefEdge",
                  "positional": "#/components/schemas/PositionalEdge"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ColumnRefEdge"
                },
                {
                  "$ref": "#/components/schemas/PositionalEdge"
                }
              ]
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeState"
            },
            "title": "Nodes",
            "type": "array"
          },
          "published_inputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedInput"
            },
            "title": "Published Inputs",
            "type": "array"
          },
          "published_outputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedOutput"
            },
            "title": "Published Outputs",
            "type": "array"
          }
        },
        "required": [
          "nodes",
          "edges"
        ],
        "title": "GraphState",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MissingPackage": {
        "description": "Package/version-level load issue for a workflow.",
        "properties": {
          "affected_nodes": {
            "items": {
              "type": "string"
            },
            "title": "Affected Nodes",
            "type": "array"
          },
          "installed_versions": {
            "items": {
              "type": "string"
            },
            "title": "Installed Versions",
            "type": "array"
          },
          "package_name": {
            "title": "Package Name",
            "type": "string"
          },
          "required_version": {
            "title": "Required Version",
            "type": "string"
          }
        },
        "required": [
          "package_name",
          "required_version"
        ],
        "title": "MissingPackage",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationResult": {
        "description": "Result of validating a graph.",
        "properties": {
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GraphValidationError"
            },
            "title": "Errors",
            "type": "array"
          },
          "node_statuses": {
            "additionalProperties": {
              "$ref": "#/components/schemas/NodeStatus"
            },
            "default": {},
            "title": "Node Statuses",
            "type": "object"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid"
        ],
        "title": "ValidationResult",
        "type": "object"
      },
      "WorkflowCreate": {
        "description": "Request body for POST /workflows.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "storage_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Path"
          }
        },
        "required": [
          "name"
        ],
        "title": "WorkflowCreate",
        "type": "object"
      },
      "WorkflowDraftCreate": {
        "description": "Request body for POST /workflow-drafts.",
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/GraphState"
          },
          "workflow_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Name"
          }
        },
        "title": "WorkflowDraftCreate",
        "type": "object"
      },
      "WorkflowDraftParameterPatch": {
        "description": "Request body for PATCH /workflow-drafts/{draft_id}/nodes/{node_id}/parameters.",
        "properties": {
          "base_revision": {
            "title": "Base Revision",
            "type": "integer"
          },
          "client_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Seq"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          }
        },
        "required": [
          "parameters",
          "base_revision"
        ],
        "title": "WorkflowDraftParameterPatch",
        "type": "object"
      },
      "WorkflowDraftResponse": {
        "description": "Draft mutation response with validation attached.",
        "properties": {
          "client_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Seq"
          },
          "dirty": {
            "default": false,
            "title": "Dirty",
            "type": "boolean"
          },
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "graph": {
            "$ref": "#/components/schemas/GraphState"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationResult"
          },
          "workflow_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Name"
          }
        },
        "required": [
          "draft_id",
          "revision",
          "graph",
          "validation"
        ],
        "title": "WorkflowDraftResponse",
        "type": "object"
      },
      "WorkflowDraftState": {
        "description": "Stored draft state returned by GET /workflow-drafts/{draft_id}.",
        "properties": {
          "client_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Seq"
          },
          "dirty": {
            "default": false,
            "title": "Dirty",
            "type": "boolean"
          },
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "graph": {
            "$ref": "#/components/schemas/GraphState"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "workflow_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Name"
          }
        },
        "required": [
          "draft_id",
          "revision",
          "graph"
        ],
        "title": "WorkflowDraftState",
        "type": "object"
      },
      "WorkflowDraftUpdate": {
        "description": "Request body for PUT /workflow-drafts/{draft_id}.",
        "properties": {
          "base_revision": {
            "title": "Base Revision",
            "type": "integer"
          },
          "client_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Seq"
          },
          "graph": {
            "$ref": "#/components/schemas/GraphState"
          }
        },
        "required": [
          "graph",
          "base_revision"
        ],
        "title": "WorkflowDraftUpdate",
        "type": "object"
      },
      "WorkflowFile": {
        "description": "Workflow file response returned to the frontend.",
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/GraphState"
          },
          "gui": {
            "additionalProperties": true,
            "title": "Gui",
            "type": "object"
          },
          "info": {
            "$ref": "#/components/schemas/WorkflowInfo"
          },
          "missing_packages": {
            "items": {
              "$ref": "#/components/schemas/MissingPackage"
            },
            "title": "Missing Packages",
            "type": "array"
          },
          "missing_tools": {
            "items": {
              "$ref": "#/components/schemas/MissingTool"
            },
            "title": "Missing Tools",
            "type": "array"
          }
        },
        "required": [
          "info",
          "graph"
        ],
        "title": "WorkflowFile",
        "type": "object"
      },
...
      "WorkflowInfo": {
        "description": "Workflow list item returned by GET /workflows.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "last_modified": {
            "title": "Last Modified",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "storage_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Path"
          }
        },
        "required": [
          "name",
          "display_name",
          "path",
          "last_modified"
        ],
        "title": "WorkflowInfo",
        "type": "object"
      },
      "WorkflowSaveBody": {
        "description": "Request body for PUT /workflows/{name}.",
        "properties": {
          "draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Id"
          },
          "graph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphState"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "title": "WorkflowSaveBody",
        "type": "object"
      },
      "WorkflowUpdate": {
        "description": "Request body for PATCH /workflows/{name}.",
        "properties": {
          "action": {
            "enum": [
              "update",
              "duplicate"
            ],
            "title": "Action",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "new_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Name"
          },
          "storage_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Path"
          }
        },
        "required": [
          "action"
        ],
        "title": "WorkflowUpdate",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "BioImageFlow Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workflow-drafts": {
      "post": {
        "operationId": "create_workflow_draft_api_v1_workflow_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDraftCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workflow Draft",
        "tags": [
          "workflow-drafts"
        ]
      }
    },
    "/api/v1/workflow-drafts/{draft_id}": {
      "get": {
        "operationId": "get_workflow_draft_api_v1_workflow_drafts__draft_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDraftState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow Draft",
        "tags": [
          "workflow-drafts"
        ]
      },
      "put": {
        "operationId": "update_workflow_draft_api_v1_workflow_drafts__draft_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDraftUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Workflow Draft",
        "tags": [
          "workflow-drafts"
        ]
      }
    },
...
    "/api/v1/workflow-drafts/{draft_id}/nodes/{node_id}/parameters": {
      "patch": {
        "operationId": "patch_workflow_draft_parameters_api_v1_workflow_drafts__draft_id__nodes__node_id__parameters_patch",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDraftParameterPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Workflow Draft Parameters",
        "tags": [
          "workflow-drafts"
        ]
      }
    },
    "/api/v1/workflow-drafts/{draft_id}/validate": {
      "post": {
        "operationId": "validate_workflow_draft_api_v1_workflow_drafts__draft_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Workflow Draft",
        "tags": [
          "workflow-drafts"
        ]
      }
    },
    "/api/v1/workflows": {
      "get": {
        "operationId": "list_workflows_api_v1_workflows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowInfo"
                  },
                  "title": "Response List Workflows Api V1 Workflows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Workflows",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "operationId": "create_workflow_api_v1_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/api/v1/workflows/{name}": {
      "delete": {
        "operationId": "delete_workflow_api_v1_workflows__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Workflow Api V1 Workflows  Name  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Workflow",
        "tags": [
          "workflows"
        ]
      },
      "get": {
        "operationId": "get_workflow_api_v1_workflows__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowFile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow",
        "tags": [
          "workflows"
        ]
      },
      "patch": {
        "operationId": "patch_workflow_api_v1_workflows__name__patch",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Workflow",
        "tags": [
          "workflows"
        ]
      },
      "put": {
        "operationId": "save_workflow_api_v1_workflows__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowSaveBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
...
  }
}